+++ src/main/java/bowling/controller/BowlingGame.java
@@ -5,7 +5,7 @@ import bowling.domain.frame.Frame;
 import bowling.domain.result.GameResult;
 import bowling.domain.result.GameResults;
 import bowling.view.InputView;
-import bowling.view.ResultView2;
+import bowling.view.ResultView;
 
 import java.util.Arrays;
 
@@ -16,6 +16,6 @@ public class BowlingGame {
         GameResult gameResult2
                 = GameResult.of(PlayerName.of(InputView.getPlayerName()), Frame.createTenFrames());
 
-        ResultView2.print(GameResults.of(Arrays.asList(gameResult1, gameResult2)));
+        ResultView.print(GameResults.of(Arrays.asList(gameResult1, gameResult2)));
     }
 }
